scrapers/requirements.txt
requests
selectolax
PyYAML
